pyproject.toml
[tool.poetry]
name = "pratapocos"
version = "0.1.0"
description = ""
authors = ["williamfoschiera <wfoschiera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
asgiref = "^3.6.0"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
django-extensions = "^3.2.1"
psycopg2 = "^2.9.6"
dj-database-url = "^2.0.0"
python-decouple = "^3.8"
django-ninja = "^0.21.0"
whitenoise = "^6.4.0"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.265"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
bandit = "^1.7.5"
pre-commit = "^3.3.1"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
old_finders = true

[tool.ruff]
line-length = 120
exclude = [
  "*/migrations/*"
]

[tool.mypy]
follow_imports = "normal"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
# disallow_untyped_defs = true  # pro futuro
# strict = true  # pro futuro

[tool.bandit]
exclude = ["*_test.py", "*.venv", "*/test_*.py", "*/conftest.py"]
recursive = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.markers]
slow="mark slow tests."

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
filterwarnings = [
    "error",
    'ignore::DeprecationWarning',
]
DJANGO_SETTINGS_MODULE = "pratapocos.pratapocos.settings"
addopts = "--reuse-db --fail-on-template-vars"  # --migrateci

[tool.coverage.run]
omit= ["tests/**"]

[tool.coverage.report]
fail_under = 50 #95
show_missing = true
exclude_lines = [
  'pragma: no cover'
]
